+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### 3.19.0(58)
+
+Models
+
+- `BasicAnimation.toCAMediaTimingFunction` is added
+- `BasicAnimation.toUIViewAnimationOptions` is added
+
 ### [3.18.4(57)](https://github.com/VakhoKontridze/VCore/releases/tag/3.18.4) — *2022 07 26*
 
 View and ViewControllers